requirements.txt
# ECG MIT-BIH Robust Classification
torch>=2.0.0
wfdb>=4.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
